Move amount proptests to amount::tests::prop
