+++ zebra-chain/src/amount/tests.rs
@@ -1,0 +1,3 @@
+//! Tests for amounts
+
+mod prop;
