+++ zebra-chain/src/value_balance/tests.rs
@@ -1,1 +1,3 @@
+//! Tests for value balances.
+
 mod prop;
