+++ zebra-chain/src/value_balance/tests/prop.rs
@@ -1,5 +1,8 @@
-use crate::{amount::*, value_balance::*};
+//! Randomised property tests for value balances.
+
 use proptest::prelude::*;
+
+use crate::{amount::*, value_balance::*};
 
 proptest! {
     #[test]
@@ -108,26 +111,4 @@ proptest! {
             prop_assert_eq!(bytes, bytes2);
         }
     }
-
-    #[test]
-    fn amount_serialization(amount in any::<Amount<NegativeAllowed>>()) {
-        zebra_test::init();
-
-        let bytes = amount.to_bytes();
-        let serialized_amount = Amount::<NegativeAllowed>::from_bytes(bytes)?;
-
-        prop_assert_eq!(amount, serialized_amount);
-    }
-
-    #[test]
-    fn amount_deserialization(bytes in any::<[u8; 8]>()) {
-        zebra_test::init();
-
-        if let Ok(deserialized) = Amount::<NegativeAllowed>::from_bytes(bytes) {
-            let bytes2 = deserialized.to_bytes();
-            prop_assert_eq!(bytes, bytes2);
-        }
-    }
-
-
 }
